Cleaned up the repo some more and updated CONTRIBUTING.md
